// src/shared/types.ts
/*
 * Copyright 2026 John Ewart <john@johnewart.net>
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export interface GitDiffFile {
  path: string;
  status: 'M' | 'A' | 'D' | 'R' | 'C' | '??';
  hunks: GitDiffHunk[];
}

export interface GitDiffHunk {
  header: string;
  lines: GitDiffLine[];
}

export interface GitDiffLine {
  type: 'context' | 'add' | 'remove';
  content: string;
  lineNumber: number;
  oldLineNumber?: number;
}

export interface DiffComment {
  id: string;
  fileId: string;
  lineNumber: number;
  text: string;
  createdAt: Date;
  updatedAt?: Date;
  isResolved?: boolean;
}